requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
dateparser==1.0.0
discord.py==1.6.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
python-binance==0.7.9
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.3
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.2.0
